+++ dl-run-tensorrt/src/export/findlinexport.h
@@ -34,4 +34,4 @@ extern "C" __declspec(dllexport) cv::Mat* _cdecl himage_to_mat(unsigned char* r,
 
 extern "C" __declspec(dllexport) int _cdecl get_vector_point_size(std::vector<app::Point>*points);
 
-extern "C" __declspec(dllexport) void _cdecl copy_vector_point(int* rows, int* cols, std::vector<app::Point>*points);
+extern "C" __declspec(dllexport) void _cdecl copy_vector_point(float* rows, float* cols, std::vector<app::Point>*points);
